Updated seed files to be more robust
