+++ DB/seeds/1_add_users.ts
@@ -7,16 +7,23 @@ export interface IUser {
   first_name: string;
   last_name: string;
   password: string;
+  role: string;
 }
+
+// Encrypted, storage-safe string for 'password'
+const encryptedPassword = "$2b$10$TaLH39x8mMOTfiYgGzv6HOKEfCLsbytzVk3BqGWmmSP3/jZ4NGIDC";
 
 export const createUser = (): IUser => ({
   email: faker.internet.email(),
   first_name: faker.name.firstName(),
   last_name: faker.name.lastName(),
-  password: faker.internet.password()
+  password: encryptedPassword,
+  role: (Math.random() >= 0.8) ? "ADMIN" : "BASIC"
 });
 
 export const seed = async (knex: Knex): Promise<void> => {
+
+  console.log("Seeding users...")
 
   // Create fake users
   const fakeUsers: Array<IUser> = [];
@@ -28,5 +35,4 @@ export const seed = async (knex: Knex): Promise<void> => {
 
   // Add users to Users table
   await knex("user_service.users").insert(fakeUsers);
-
 }
